+++ NEWS.md
@@ -12,6 +12,7 @@
 - Added more indexes to `list_supported_indexes()`.
 - Added `download_data_huggingface()` and `get_available_huggingface_files()`. and support for `type = "hf_high_frequency_sp500"`.
 - Deprecated `type` parameter in favor of `domain` and `dataset`.
+- Added `detail` parameter to `estimate_fama_macbeth()` to include average `n_obs`, `r_squared`, and `adj_r_squared`.
 - Removed lower bound of excess returns for CRSP data. 
 - Removed `add_lag_columns()` in favor of `add_lagged_columns()`.
 - Added domain `"tidyfinance"` with datasets `"high_frequency_sp500"` and `"factor_library"`.